+++ test/Scripts.Integration.Test/Editor/Builder.cs
@@ -16,20 +16,20 @@ public class Builder
         // Make sure the configuration is right.
         EditorUserBuildSettings.selectedBuildTargetGroup = group;
         PlayerSettings.SetScriptingBackend(group, ScriptingImplementation.IL2CPP);
-
-        if (args.ContainsKey("sentryOptions.configure"))
-        {
-            SetupSentryOptions(args);
-        }
+        DisableUnityAudio();
 
         var buildPlayerOptions = new BuildPlayerOptions
         {
-            scenes = new[] { "Assets/Scenes/1_Bugfarm.unity" },
             locationPathName = args["buildPath"],
             target = target,
             targetGroup = group,
             options = BuildOptions.StrictMode,
         };
+
+        if(File.Exists("Assets/Scenes/SmokeTest.unity"))
+        {
+            buildPlayerOptions.scenes = new[] { "Assets/Scenes/SmokeTest.unity" };
+        }
 
         var report = BuildPipeline.BuildPlayer(buildPlayerOptions);
         var summary = report.summary;
@@ -62,60 +62,12 @@ public class Builder
             Debug.Log($"Build succeeded with {summary.totalWarnings} warning{(summary.totalWarnings > 1 ? "s" : "")}.");
         }
     }
-
     public static void BuildWindowsIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.StandaloneWindows64, BuildTargetGroup.Standalone);
     public static void BuildMacIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.StandaloneOSX, BuildTargetGroup.Standalone);
     public static void BuildLinuxIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.StandaloneLinux64, BuildTargetGroup.Standalone);
     public static void BuildAndroidIl2CPPPlayer() => BuildIl2CPPPlayer(BuildTarget.Android, BuildTargetGroup.Android);
     public static void BuildIOSPlayer() => BuildIl2CPPPlayer(BuildTarget.iOS, BuildTargetGroup.iOS);
     public static void BuildWebGLPlayer() => BuildIl2CPPPlayer(BuildTarget.WebGL, BuildTargetGroup.WebGL);
-
-    private static void SetupSentryOptions(Dictionary<string, string> args)
-    {
-        Debug.LogFormat(LogType.Log, LogOption.NoStacktrace, null, "SetupSentryOptions: Invoking SentryOptions");
-
-        if (!EditorApplication.ExecuteMenuItem("Tools/Sentry"))
-        {
-            throw new Exception("SetupSentryOptions: Menu item Tools -> Sentry was not found.");
-        }
-
-        var sentryWindowType = AppDomain.CurrentDomain.GetAssemblies()
-            ?.FirstOrDefault(assembly => assembly.FullName.StartsWith("Sentry.Unity.Editor,"))
-            ?.GetTypes()?.FirstOrDefault(type => type.FullName == "Sentry.Unity.Editor.SentryWindow");
-        if (sentryWindowType is null)
-        {
-            foreach (var asm in AppDomain.CurrentDomain.GetAssemblies().Where(name => name.FullName.Contains("Sentry")))
-            {
-                foreach (var type in asm.GetTypes())
-                {
-                    Debug.LogFormat(LogType.Log, LogOption.NoStacktrace, null, "SetupSentryOptions: Asm {0} Type {1}", asm.FullName, type.FullName);
-                }
-            }
-            throw new EntryPointNotFoundException("SetupSentryOptions: Type SentryWindow not found");
-        }
-
-        var optionsWindow = EditorWindow.GetWindow(sentryWindowType);
-        var options = optionsWindow.GetType().GetProperty("Options").GetValue(optionsWindow);
-
-        if (options is null)
-        {
-            throw new EntryPointNotFoundException("SetupSentryOptions: Method SentryOptions not found");
-        }
-        Debug.LogFormat(LogType.Log, LogOption.NoStacktrace, null, "SetupSentryOptions: Found SentryOptions");
-
-        var dsn = args["sentryOptions.Dsn"];
-        if (dsn != null)
-        {
-            Debug.LogFormat(LogType.Log, LogOption.NoStacktrace, null,
-                "SetupSentryOptions: Configuring Dsn to {0}", dsn);
-
-            var dnsPropertyInfo = options.GetType().GetProperty("Dsn");
-            dnsPropertyInfo.SetValue(options, dsn, null);
-        }
-
-        optionsWindow.Close();
-        Debug.LogFormat(LogType.Log, LogOption.NoStacktrace, null, "SetupSentryOptions: Sentry options Configured");
-    }
 
     public static Dictionary<string, string> ParseCommandLineArguments()
     {
@@ -145,5 +97,15 @@ public class Builder
         {
             throw new Exception("No valid '-buildPath' has been provided.");
         }
+    }
+
+    // Audio created issues, especially for iOS simulator so we disable it.
+    private static void DisableUnityAudio()
+    {
+        var audioManager = AssetDatabase.LoadAllAssetsAtPath("ProjectSettings/AudioManager.asset")[0];
+        var serializedManager = new SerializedObject(audioManager);
+        var prop = serializedManager.FindProperty("m_DisableAudio");
+        prop.boolValue = true;
+        serializedManager.ApplyModifiedProperties();
     }
 }