+++ test/Scripts.Integration.Test/Scripts/CppPlugin.cpp
@@ -1,0 +1,6 @@
+extern "C" {
+void throw_cpp()
+{
+    throw "exception";
+}
+}