+++ test/Scripts.Integration.Test/integration-test.ps1
@@ -1,7 +1,9 @@
-﻿param($path)
+﻿param(
+    [string] $UnityPath
+)
 
-.\test/Scripts.Integration.Test/integration-create-project.ps1 "$path"
-.\test/Scripts.Integration.Test/integration-build-project.ps1 "$path"
-.\test/Scripts.Integration.Test/integration-update-sentry.ps1 "$path"
-.\test/Scripts.Integration.Test/integration-build-project.ps1 "$path"
-.\test/Scripts.Integration.Test/integration-run-smoke-test.ps1
+.\test/Scripts.Integration.Test/create-project.ps1 "$UnityPath"
+.\test/Scripts.Integration.Test/build-project.ps1 "$UnityPath"
+.\test/Scripts.Integration.Test/update-sentry.ps1 "$UnityPath"
+.\test/Scripts.Integration.Test/build-project.ps1 "$UnityPath"
+.\test/Scripts.Integration.Test/run-smoke-test.ps1 -Smoke